Updates to tenting foot and thumb cluster
This update removes most, if not all, uses of the "up to face" padding tool from the Tenting Foot. It turns out "up to face" is still pretty buggy in FreeCAD. You can still adjust the tenting angle of the keyboard, but you cannot change the location of Tenting foot's planes. In the future, the plane locations will be set by the spreadsheet and this will not be an issue. On the plus side, Tenting foot A and Tenting foot B are now identical, so if you accidentally print the same one twice it will still work.

The thumb cluster now has a horizontal screw slot instead of 2 screw holes. You should now be able to slide the cluster left and right as well as twist it. In reality, the amount of adjustability is very small because column 3 gets in the way; you can duplicate column 4 and lower its height if you want to fix this.