Minor update to personal layout
This is just some minor improvements to my personal Chimera layout. I'm not satisfied with it yet; I think I'll need a dedicated mouse input layer.
If you're getting notifications on this repo, I do have big plans for a V2.0 and QMK, but it might be a few months before I can commit real time to this project again. It depends on my school workload.